app_state: return error codes when processing
Return any error codes when processing the desired state. Previously,
there was a possibility that error codes would be generated but not
returned. This commit returns codes where possible.

Note that when updating the desired and actual endpoints, it is possible
that both will generate an error code but only the code from the desired
state operation will be returned. This is a known limitation to make
sure both operations are attempted.

Signed-off-by: Mike Szczys <mike@golioth.io>