properly size stack values
The system work queue stack size is set in the prj.conf, but then half of
that size is selected in both of the board conf files. Even with the
reduced size the utilization is still only around 60% so this value is now
used in prj.conf and has been removed from board files.

The main stack size is also just over 1k so the value has been reduced from
4k to 2k in this commit.

Signed-off-by: Mike Szczys <mike@golioth.io>