settings: wake main thread after updating delay
When the LOOP_DELAY_S setting is updated, we need to wake the main thread
which is currently sleeping based on the old delay value. This means you
will have a sensor reading every time a new setting is received, but the
point is for new setting to immediately take effect.

Signed-off-by: Mike Szczys <mike@golioth.io>