Fix RPC registration error handling
Previously it was possible for some of the RPC registrations to fail,
but as long as the last one succeeded, app_rpc_register would
incorrectly return success. Since the return value of app_rpc_register
is not being checked anywhere, this just removes the return value (an
error will still be printed when any of the registrations fail).

Signed-off-by: Chris Wilson <chris@cgnd.dev>