app_sensors: use CBOR instead of JSON
The default Golioth Pipeline for a new project applies to CBOR data and not
JSON data. This updates the data type to use CBOR.

NOTE: By default the pipeline will not include the path with the data. The
RD Template sends to the "sensor" path, but LightDB data will not be shown
under that path unless an inject transform is added.

Signed-off-by: Mike Szczys <mike@golioth.io>