app_sensors: only stream sensor data if connected
When using LTE, we may start fetching sensor readings before a connection
is established with Golioth. Readings can be shown on a local display
during this period, but attempting to send to Golioth will fail. This
updates the sensor reading streaming, which is just a counter in this case,
to only attempt to stream readings to Golioth when connected. This mirrors
the behavior of the battery monitor.

Signed-off-by: Daniel Mangum <georgedanielmangum@gmail.com>