README: remove hardware variations sections
This section was primarily used for build commands for the Golioth custom
boards. In almost all cases, this is not needed by users. Only in cases
where we have sent loaner units is it relevant, and in those cases it's
more likely the user needs instructions on using precompiled binaries from
the latest release.

Future work will add these instructions to the Projects pages.

Signed-off-by: Mike Szczys <mike@golioth.io>