battery monitor: add support for MAX17262 fuel gauge
Add support for reading the Elixir battery voltage and percentage using the
MAX17262 fuel gauge.

By default the library will now use the fuel gauge if a devicetree node for
the MAX17262 is enabled. If not, the library will use a voltage divider
that has the devicetree path `vbatt`.

The library should only be used when one of the two above cases is
present for a board.

Signed-off-by: Mike Szczys <mike@golioth.io>