Add support for NXP mimxrt1024_evk
This was an experiment to gauge the effort involved in building for non-NCS
targets. The effort is minimal, as reflected in this PR.

However it is not clear if we can merge this support as the need for more
than one manifest file to support the changeover between NCS and Zephyr so
there are currently no plans to merge this commit into main.

Signed-off-by: Mike Szczys <mike@golioth.io>